+++ tests/test_to_thread.py
@@ -3,10 +3,12 @@ import sys
 import threading
 import time
 from concurrent.futures import Future
+from contextvars import ContextVar
 from functools import partial
 from typing import Any, List, NoReturn, Optional
 
 import pytest
+import sniffio
 
 import anyio.to_thread
 from anyio import (
@@ -130,6 +132,17 @@ async def test_cancel_wait_on_thread() -> None:
 
     await to_thread.run_sync(event.set)
     assert future.result(1)
+
+
+async def test_contextvar_propagation() -> None:
+    var = ContextVar('var', default=1)
+    var.set(6)
+    assert await to_thread.run_sync(var.get) == 6
+
+
+async def test_asynclib_detection() -> None:
+    with pytest.raises(sniffio.AsyncLibraryNotFoundError):
+        await to_thread.run_sync(sniffio.current_async_library)
 
 
 @pytest.mark.parametrize('anyio_backend', ['asyncio'])